// src/enums/user.ts
/* eslint-disable no-unused-vars */
export enum ENUM_USER_ROLE {
  ADMIN = "ADMIN",
  TRAINER = "TRAINER",
  TRAINEE = "TRAINEE",
}
export enum ENUM_GENDER_USER {
  MALE = "MALE",
  FEMALE = "FEMALE",
  OTHER = "OTHER",
}

export type Role = "ADMIN" | "TRAINER" | "TRAINEE";
export type Gender_User = "MALE" | "FEMALE" | "OTHER";